+++ Assets/Scripts/UI/FirebaseAuthUI.cs
@@ -1,5 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using TMPro;
 using Game.DB;
